refactor: remove thin wrapper components around GameMenuCard (#205)
Delete SimonMenuScreen, MathRaceMenuScreen, ColorTapMenuScreen,
WordScrambleMenuScreen, WordBuilderMenuScreen, WhackMenuScreen,
TrueFalseMenuScreen, EmojiMathMenuScreen, BalloonMenuScreen — each
was a one-liner that only forwarded hardcoded theme props to GameMenuCard.
Use GameMenuCard directly in the parent game component instead.
