+++ gamesformykids/app/games/word-builder/WordBuilderGame.tsx
@@ -1,7 +1,7 @@
 ﻿'use client';
 
 import { useWordBuilderGame } from './useWordBuilderGame';
-import WordBuilderMenuScreen from './components/WordBuilderMenuScreen';
+import GameMenuCard from '@/components/game/shared/GameMenuCard';
 import WordBuilderQuestion from './components/WordBuilderQuestion';
 import WordBuilderResultScreen from './components/WordBuilderResultScreen';
 
@@ -11,7 +11,17 @@ export default function WordBuilderGame() {
     startGame, pressLetter, clearTyped, next, restart,
   } = useWordBuilderGame();
 
-  if (phase === 'menu') return <WordBuilderMenuScreen onStart={startGame} />;
+  if (phase === 'menu') return (
+    <GameMenuCard
+      emoji="🔤"
+      title="בניית מילים"
+      description="סדר את האותיות ובנה את המילה הנכונה!"
+      gradientClass="from-orange-50 to-amber-100"
+      buttonClass="from-orange-500 to-amber-500"
+      onStart={startGame}
+      startLabel="🚀 התחל!"
+    />
+  );
 
   if (phase === 'playing' && current) return (
     <WordBuilderQuestion
